website/requirements.txt
aldryn-apphooks-config==0.4.1
aldryn-boilerplates==0.7.7
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-newsblog==2.1.1
aldryn-people==2.1.0
aldryn-translation-tools==0.2.1
Babel==2.6.0
backport-collections==0.1
certifi==2018.8.13
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.15
django-appconf==1.0.2
django-appdata==0.2.0
django-classy-tags==0.8.0
django-cms==3.5.2
django-filemanager==0.0.5
django-filer==1.3.2
django-formtools==2.1
django-mptt==0.8.7
django-parler==1.9.2
django-phonenumber-field==1.3.0
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-sortedm2m==1.5.0
django-taggit==0.22.2
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
easy-thumbnails==2.5
html5lib==0.9999999
idna==2.7
lxml==4.2.4
phonenumbers==8.9.10
phonenumberslite==8.9.10
Pillow==5.2.0
python-dateutil==2.7.3
python-slugify==1.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.23
YURL==0.13
